requirements.txt
accelerate
git+https://github.com/descriptinc/audiotools
transformers
tokenizers
cn2an
ffmpeg-python
Cython
g2p-en
jieba
json5
keras
numba
numpy
pandas
matplotlib
munch
opencv-python
tensorboard
librosa
safetensors
modelscope
omegaconf
sentencepiece
tqdm
textstat
torch-complex
pynini==2.1.6; platform_system!="Windows"
WeTextProcessing>=1.0.3; platform_system!="Windows"

WeTextProcessing; platform_machine != "Darwin"
wetext; platform_system == "Darwin"

# importlib_resources
# pynini==2.1.6.post1
# WeTextProcessing>=1.0.4
# deepspeed # Use it to accelerate model inference